paper/requirements.txt
astropy==5.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
corner==2.2.1
cycler==0.11.0
Cython==0.29.30
dynesty==1.2.3
emcee==3.1.2
fonttools==4.34.4
h5py==3.7.0
halotools==0.7
idna==3.3
joblib==1.1.0
kepler.py==0.0.6
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pocomc==0.1.1
pybind11==2.10.0
pyerfa==2.0.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
setuptools-scm==7.0.5
six==1.16.0
soupsieve==2.3.2.post1
tabcorr==1.0.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.0
tqdm==4.64.0
typing-extensions==4.3.0
ultranest==3.4.6
urllib3==1.26.10
